pyproject.toml
[project]
name = "doc-flow"
version = "0.1.0"
description = "Generic document flow processor for state tracking in markdown front matter"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.1",
    "pyyaml>=6.0.3",
]

[project.scripts]
doc-flow = "doc_flow.cli:main"

[tool.uv]
package = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
]

[tool.ruff.format]
quote-style = "double"
